requirements.txt
flask
torch
torchvision
timm
pillow
numpy
requests
